+++ src/components/common/layouts/wrapped-content-layout.tsx
@@ -1,6 +1,6 @@
 import styles from '@/styles/components/common/layouts/wrapped-content-layout.module.scss';
 import { CSSProperties, ReactNode } from 'react';
-import { Header } from '@/src/components/common/header';
+import { Header } from '@/src/components/common/header/header';
 import { Footer } from '@/src/components/common/footer';
 
 interface WrappedContentLayoutProps {