+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2025-04-15
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Sidebar components.
+- [Marek Kurańda](https://github.com/mjkuranda): `useAppView` hook for detecting which device is used.
+- [Marek Kurańda](https://github.com/mjkuranda): `useUser` hook for managing a user account.
+- [Marek Kurańda](https://github.com/mjkuranda): `useHeaderLinks` hook for getting all links needed to be within the sidebar.
+- [Marek Kurańda](https://github.com/mjkuranda): `Nav` and `AccountDropdown` components.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Depending on used device, header renders sidebar or nav component.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): `User` component.
+
 ## [2.3.0] - 2025-03-15
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Mixin for different type of devices: `mobile`,`tablet` and `desktop`.